+++ content/js/portal/isolated.js
@@ -16,49 +16,44 @@ class Watcher {
   }
 }
 
-// class VisibilityRestorer extends Watcher {
-//   constructor() {
-//     super();
-//     this.observer = new MutationObserver(this.detectTargetElement.bind(this));
-//     this.inputMap = {};
-//     this.propName = 'resourceVisibility';
-//     this.SELECTOR_TARGET_ELEMENT = 'section:last-of-type .ext-hubs-artbrowse-filter-showall .azc-validatableControl-none.azc-text-label';
-//   }
-//   async updateTargetElement(inputEvent) {
-//     inputEvent.stopPropagation();
-//     const view = document.location.hash.replace(/^[\S\s]*\/subscriptions/, '/subscriptions') || '';
-//     if (!view) return;
-//     const input = inputEvent.currentTarget.querySelector('input[type="checkbox"]');
-//     if (!input) return;
+class CloudShellOpener extends Watcher {
+  constructor() {
+    super();
+    this.observer = new MutationObserver(this.detectTargetElement.bind(this));
+    this.targetElement = null;
+    this.csWindow = null;
+    this.clickHandler = (function (event) {
+      if (!window.opener && document.querySelector('#consoleFrameId') && !document.querySelector('.fxs-console').classList.contains('fxs-display-none')) return;
+      event.stopPropagation();
+      event.preventDefault();
+      this.openCloudShell();
+    }).bind(this);
+  }
 
-//     if (input.checked) this.options[this.propName][view] = input.checked;
-//     else delete this.options[this.propName][view];
+  openCloudShell() {
+    const cloudShellUrl = 'https://portal.azure.com/#cloudshell/';
+    window.csPopup = this.csWindow = window.open(cloudShellUrl, 'csInNewTab');
+  }
 
-//     await chrome.storage.local.set({
-//       "visibilityRestorer": {
-//         status: true,
-//         options: this.options
-//       }
-//     });
-//   }
-//   detectTargetElement() {
-//     const targetElements = [...document.querySelectorAll(this.SELECTOR_TARGET_ELEMENT)];
-//     const view = document.location.hash.replace(/^[\S\s]*\/subscriptions/, '/subscriptions') || '';
-//     if (targetElements.length === 0 || !view || this.inputMap[view] == targetElements[0]) return;
-//     this.inputMap[view] = targetElements[0];
-//     if (this.options[this.propName][view]) this.inputMap[view].click();
-//     this.inputMap[view].addEventListener('click', this.updateTargetElement.bind(this));
-//   }
-//   startWatching(options) {
-//     this.options = options || { [this.propName]: {} };
-//     this.detectTargetElement();
-//     this.observer.observe(document, { childList: true, subtree: true });
-//   }
+  detectTargetElement() {
+    const targetElements = [...document.querySelectorAll('[title="Cloud Shell"][aria-label="Cloud Shell"]')];
+    if (targetElements.length === 0 || this.targetElement) return;
+    this.targetElement = targetElements[0];
+    this.targetElement.addEventListener('click', this.clickHandler, { capture: true });
+  }
 
-//   stopWatching() {
-//     super.stopWatching();
-//   }
-// }
+  startWatching(options) {
+    this.options = options;
+    this.detectTargetElement();
+    this.observer.observe(document, { childList: true, subtree: true });
+  }
+
+  stopWatching() {
+    super.stopWatching();
+    this.targetElement && this.targetElement.removeEventListener('click', this.clickHandler, { capture: true });
+    this.targetElement = null
+  }
+}
 
 class FilterRestorer extends Watcher {
   constructor() {
@@ -107,7 +102,7 @@ class AdvancedCopy extends Watcher {
   constructor() {
     super();
     this.messageQueue = [];
-    this.re = /(\/subscriptions\/[0-9a-f]{8}(?:-[0-9a-f]{4}){4}[0-9a-f]{8}\/resourceGroups\/([^/]+)\/providers\/([^/]+)\/([^/]+)\/([^/]+))/i;
+    this.re = /(\/subscriptions\/[0-9a-f]{8}(?:-[0-9a-f]{4}){4}[0-9a-f]{8}\/resourceGroups\/([^/]+)\/providers\/([^/]+)(?:\/([^/]+)\/([^/]+))+)/i;
     this.cache = {};
     this.menus = [{
       title: 'Resource name',
@@ -1103,6 +1098,7 @@ const storeAccessToken = async () => {
     _watchers['advancedCopy'] = new AdvancedCopy();
     _watchers['filterRestorer'] = new FilterRestorer();
     _watchers['contextMenuUpdater'] = new ContextMenuUpdater();
+    _watchers['cloudShellOpener'] = new CloudShellOpener();
     // _watchers['visibilityRestorer'] = new VisibilityRestorer();
 
     const init = async (changes) => {
@@ -1112,6 +1108,7 @@ const storeAccessToken = async () => {
       })(changes, _watchers);
 
       watcherStatus['contextMenuUpdater'] = { status: true };
+      // watcherStatus['cloudShellOpener'] = { status: true };
       Object.keys(watcherStatus).forEach(w => {
         if (!watcherStatus[w] || !_watchers[w]) return;
         if (watcherStatus[w].status) _watchers[w].startWatching(watcherStatus[w].options);