+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Release History
+## [1.1.2](https://github.com/horihiro/TweakIt-for-Azure-ChromeExtension/releases/tag/1.1.2)
+### New feature(s)
+  - Cloud Shell: support opening Cloud Shell in a new tab
+
+### Bug Fix(es)
+  - Advanced Copy: copying resource ids for sub-resources
+  - Cloud Shell: enhance storage handling for cloud shell options in settings
+  - Cloud Shell: clear default startup commands and improve socket handling
+
 ## [1.1.1](https://github.com/horihiro/TweakIt-for-Azure-ChromeExtension/releases/tag/1.1.1)
 ### Improve feature(s)
   - Options / UI: add default css for resource group list decoration